Fix issue where adding game to library removes all other platforms from wishlist. Add overall wishlist view to More tab.

+++ Backlogger/View Controllers/GameDetailsViewController.swift
@@ -1343,7 +1343,7 @@ extension GameDetailsViewController: ConsoleSelectionTableViewControllerDelegate
             
             if consoles.count > 0 {
                 for game in (self._gameField?.ownedGames)! {
-                    if !consoles.contains(game.platform!) {
+                    if !consoles.contains(game.platform!) && game.inLibrary {
                         game.delete()
                     } else {
                         currentPlatformList.append(game.platform!)
